a ver si funca ome...

+++ ciclo4amongo/src/main/java/co/edu/unbosque/ciclo4amongo/repository/ProductRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import co.edu.unbosque.ciclo4amongo.model.Product;
 //Este es un comentario re chimbo
-public interface ProductRepository extends MongoRepository < Product, Long > {
+public interface ProductRepository extends MongoRepository < Product, Integer > {
 
 }
